More Shubibinman
I spent WAY too long on this, much more than I should have. But hey! It
turns out that the 2p controls in PCEHawk are in a really obtuse place
on the emulator. Welp.

Anyways, I did a 2p version and at this point i'm rather tired of this
game. So here you go.