BizHawk 1.11.9.1 Test
Happy Holdays to all that view this GitHub besides me!

Apparently, There was a new version of BizHawk that was released. So I
decided to give it a shot to check out what this new version changes.
The result was two runs of games I had previously done.

The first one is Psycho Fox, which seems to be faster than my other run
by 15 frames. This might be the first time I really try and grind for
frames of improvement. There seems to be a bit of emulation differences
here, but it doesn't seems to make much of a difference in the run.

The other run is of Vigilante, which at first glance seems to be MUCH
faster than my other run. I was never entirely too happy with the
quality of my current run, so I am glad that I did this run. It's also
kinda interesting to see how my runs improve over time with the skills
that i've learned over the years of speed running.

The rest of this commit is just various additions that werent previously
there (movie files, watch files) as well as some edits for
clarification.

I wish you all the best during the holidays and a happy new year!