Toki
Did a version that was 32 frames faster than the previous movie. Didn't
bother making it a different version since it was so close to the end.

Okay, i'm gonna be honest. It seems as though BizHawk is kinda in a
state of flux in terms of emulatin accuracy. So i'm probably going to
stop doing assorted TAS stuff until the emulation is a little
more...consistent. Or at least at a point where core changes aren't so
frequent.