Update: 7/23/17
Welp, it's certainly been a while since I committed something to this
repro.

Mickey Mania: So about that Mickey Mania Test Run that was in the
Unfinished WIPs folder. I ended up finishing it. Granted, it's nowhere
NEAR as polished as my other work, I just wanted to see what a TAS would
look like with all the new strats.

McDonald's TLA: This one is kinda intersting. So apparently there seems
to be an entire community of people running this game now, complete with
a person by the name "node" doing some TAS stuff with new strats. As an
act of kindness I suppose, I made this short movie that should have some
newely updated strats for him as a guide in the future. At some point,I
might decide to do an entirely new Expert TAS. However, that entirely
depends on whether or not I muster up the motivation to go through with
making one. Either way, here it is in all of its glory.