McDonald's TLA
I haven't done a commit in a long time and ho boy, do I got some stuff
for you guys.

McDonald's TLA - So this is a new Expert TAS I created with node1729, a
prominent member of the McDonald's speedrunning community (I can't
believe those were words that were put together). It ended up being over
a minute and thirty seconds faster than the TAS I did in 2009. Needless
to say, I'm quite happy how it turned out. Were still very much checking
over improvements so this might not be the finalized TAS. Once
everything is done, node was planning on submitting it to TASvideos, so
I guess take that for what it's worth. :P

Kncukles Chaotix - I tried out the first Tower level and ended up
zipping to the left. I didn't even know that was a thing. lol