4/19/18 Update!
Wiz 'n' Liz - Ok, So apparently the Apprentice TAS was actually not the
complete TAS that I did, but rather an attempt I made on Super Wizard.
That file should be replaced. On top of that, I made an actual attempt
at Super Wizard, So have fun!

Granada - Tried like 30 seconds of this. I thought this would be 60
FPS...I guess not.

Street Racer - Ok, you don't undertand. This game is like fast...like
really fast. So I tried messing around witht he first track and got a
1:38. I have no idea how the mechanics work, so if this looks fast
enough for you feel free to try and optimize it. At the very least, this
does seem like it would make for a great speedgame in the future. Hmm...