Lets get dirty (dwarves).
Three Dirty Dwarves - So some of you may know that I was holding off on
doing any more TAS bits with BizHawk since the accuracy of the cores has
been constantly changing. This is a little different.

Three Dirty Dwarves is a game that i've been eyeing for a while as a
possible new speedgame. In order to be better prepared for that fateful
day, I wanted to try and make a TAS on Hard to get an idea of what the
time/route would be like. Unfortunately, I just don't have the time to
do a large-scale TAS like this right now. Maybe in the future when my
living conditions improve, I might try to tacle this in the future. For
now i'll leave this here in case others want to take a look.