New Update. Happy 2023!
Long time no see!

So BizHawk recently (or I guess not so recently) came out with an Arcade core and I just had to take a look. When I announced my retirement from TASing back in 2012 that caused all of the controversy/drama surrounding my SoR2 TAS, I generally started to have differing thoughts on the state of emulation which was doubly true as I was getting back to my routes with real-time speedrunning. I felt that if more TASes were to be made, then there should be more of an effort to make sure the emulation was as accurate to original hardware as possible in order to have some form of parity to RTA runners. And given that the state of emulation was constantly changing, it didn't really motivate me to do much of anything regarding emulators. The only exception I would give would be with Arcade emulation since that's a platform you can't really be 100% accurate on (for obvious reasons).

Now that BizHawk supports Arcade emulation, I finally have more of a chance to mess around with arcade games in a way that I wouldn't have been able to before. For contrast, the only other emulators available at the time were FBA and MAME which while serviceable weren't exactly what I was looking for. Needless to say, this is a HUGE deal and I can't wait to see what fun stuff comes out of this new paradigm.

New to this update:
*Mystic Warriors "WIP"
*Violent Storm "WIP"
* Biomechanical Toy "Improvement" (Was originally an fbm file but was redone in BizHawk)